edit-in-place (our first modal component)
This is an example of how to use the edit-in-place modal to create
an editable item.

We use a bunch of semantic HTML to make the JavaScript fairly
reusable, though the strategy to save a particular key/value pair
might change.  (Keep an eye on later commits for a customization
strategy.)

Functionality can be observed by going into the Head Items section
and clicking on an attribute name.